+++ jkube-kit/generator/api/src/main/java/org/eclipse/jkube/generator/api/support/BaseGenerator.java
@@ -14,10 +14,7 @@
 package org.eclipse.jkube.generator.api.support;
 
 import java.io.IOException;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.time.format.FormatStyle;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -292,4 +289,5 @@ public abstract class BaseGenerator implements Generator {
             buildBuilder.labels(labels);
         }
     }
+
 }