+++ jkube-kit/generator/java-exec/src/test/java/org/eclipse/jkube/generator/javaexec/JavaExecGeneratorTest.java
@@ -14,6 +14,8 @@
 package org.eclipse.jkube.generator.javaexec;
 
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.LocalDate;
@@ -23,35 +25,52 @@ import java.util.List;
 import java.util.Properties;
 
 import org.eclipse.jkube.generator.api.GeneratorContext;
+import org.assertj.core.api.InstanceOfAssertFactories;
 import org.eclipse.jkube.kit.common.Assembly;
 import org.eclipse.jkube.kit.common.AssemblyConfiguration;
 import org.eclipse.jkube.kit.common.JavaProject;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.common.Plugin;
-import org.assertj.core.api.InstanceOfAssertFactories;
 import org.eclipse.jkube.kit.config.image.ImageConfiguration;
 import org.eclipse.jkube.kit.config.image.build.BuildConfiguration;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
+import org.mockito.MockedConstruction;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.mockConstruction;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 class JavaExecGeneratorTest {
   private JavaProject project;
   private GeneratorContext generatorContext;
   private Properties properties;
 
+  @TempDir
+  Path temporaryFolder;
 
   @BeforeEach
-  void setUp() {
+  void setUp() throws IOException {
     properties = new Properties();
+    final File targetDir = Files.createDirectory(temporaryFolder.resolve("target")).toFile();
     project = JavaProject.builder()
       .properties(properties)
+      .baseDirectory(temporaryFolder.toFile())
+      .buildDirectory(targetDir)
+      .buildPackageDirectory(targetDir)
+      .outputDirectory(targetDir)
       .version("0.0.1")
       .build();
     generatorContext = GeneratorContext.builder()
-      .logger(new KitLogger.SilentLogger())
+      .logger(spy(new KitLogger.SilentLogger()))
       .project(project)
       .build();
   }
@@ -123,6 +142,43 @@ class JavaExecGeneratorTest {
             tuple(new File("src/main/jkube-includes/bin"), Collections.emptyList()),
             tuple(new File("fatjar-simple"), Collections.singletonList("fat.jar"))
         );
+  }
+
+
+  @Test
+  @DisplayName("createAssembly, logs info if the final output artifact is same as previous")
+  void createAssembly_logsFinalArtifactSameAsPrevious() throws IOException {
+    final Path fatJar = Files.createFile(project.getBuildDirectory().toPath().resolve("fat.jar"));
+    try (MockedConstruction<FatJarDetector> fatJarDetector = mockConstruction(FatJarDetector.class,
+      (mock, ctx) -> {
+        FatJarDetector.Result mockedResult = mock(FatJarDetector.Result.class);
+        when(mockedResult.getArchiveFile()).thenReturn(fatJar.toFile());
+        when(mock.scan()).thenReturn(mockedResult);
+      })) {
+      JavaExecGenerator generator = new JavaExecGenerator(generatorContext);
+      generator.createAssembly();
+      generator.createAssembly(); // calling twice to simulate that the final output artifacts were not changed
+      // Then
+      verify(generatorContext.getLogger())
+        .info("Final output artifact file was not rebuilt since last build. " +
+          "HINT: try to compile and package your application prior to running the container image build task.");
+    }
+  }
+
+  @Test
+  @DisplayName("createAssembly, doesn't log info if artifact is recently built")
+  void createAssembly_doesntLogWarnings() throws IOException {
+    final Path fatJar = Files.createFile(project.getBuildDirectory().toPath().resolve("fat.jar"));
+    try (MockedConstruction<FatJarDetector> fatJarDetector = mockConstruction(FatJarDetector.class,
+      (mock, ctx) -> {
+        FatJarDetector.Result mockedResult = mock(FatJarDetector.Result.class);
+        when(mockedResult.getArchiveFile()).thenReturn(fatJar.toFile());
+        when(mock.scan()).thenReturn(mockedResult);
+      })) {
+      new JavaExecGenerator(generatorContext).createAssembly();
+      verify(generatorContext.getLogger(), times(0)).info(anyString());
+      verify(generatorContext.getLogger(), times(0)).warn(anyString());
+    }
   }
 
   @Test