+++ jkube-kit/jkube-kit-quarkus/src/test/java/org/eclipse/jkube/quarkus/generator/QuarkusGeneratorTest.java
@@ -192,7 +192,7 @@ class QuarkusGeneratorTest {
       // Then
       assertThat(result).singleElement()
         .extracting("buildConfiguration.from").asString()
-        .startsWith("registry.access.redhat.com/ubi8/ubi-minimal:");
+        .startsWith("registry.access.redhat.com/ubi9/ubi-minimal:");
     }
 
     @Test
@@ -494,7 +494,7 @@ class QuarkusGeneratorTest {
       // When & Then
       assertThatIllegalStateException()
         .isThrownBy(() -> qg.customize(configs, false))
-        .withMessageContaining("The quarkus-app directory required in Quarkus Fast Jar mode was not found");
+        .withMessageContaining("The quarkus-app directory required in Quarkus Fast Jar mode was not found. HINT: try to compile and package your application prior to running the container image build task.");
     }
   }